Create processTotext.cpp
A simple application that saves date and time to a text file, and a list of processes running in the background. The program also scans all processes every second, displays the unix time and its PID. If you run another copy of the application, the previous one will be closed.